refactor(0X-Guide): rename 05 files move incompletes to debug
