+++ 05-Callbacks-Iterators/02-startingCapsOnly.js
@@ -4,9 +4,13 @@ import { uppers } from "./data/startingCapsOnly.data";
  * Define the function `startingCapsOnly`.
  * `startingCapsOnly` accepts an array of Strings; the array is always .length >= 1
  * `startingCapsOnly` returns a new array made of any words that begin with a capital letter.
+ *
+ * You have been given an array, `uppers`, of all uppercase letters to use for this problem.
+ *
  * Use `.filter`!
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter
  *
+ * @category 05 - Callbacks
  * @function startingCapsOnly
  * @param {String[]} words
  * @returns {String[]}
