+++ debug/04-stringAssembler.js
