+++ debug/05-stringSorter.test.js
@@ -15,7 +15,7 @@ import {
   scsResult,
   noCapStart,
   ncsResult,
-} from "../data/stringSorter.data";
+} from "./data/stringSorter.data";
 import { stringSorter } from "../05-stringSorter";
 
 describe.only("#5: stringSorter", () => {
